Fixes wrongly extracted time components
As written on discord the hours / minutes and days are wrongly extracted. They have to be calculated from midnight and also with the correct division or modulus. I am testing days and months to right now but the hours are definitely wrong.

You can test it by using a value of 851512. It shows 56 hours where it should be like only 20 hours. I am available in Discord if questions remain :)